+++ CHANGELOG.md
@@ -16,6 +16,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 -->
+## 1.13.4 - 2024-06-24
+
+### Bugs fixes 🐛
+
+* fix incompatible option type for Python < 3.10 and restore Python 3.8 and 3.9 by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/304> and @pawamoy in <https://github.com/mkdocs/mkdocs/issues/3753#issuecomment-2187180187>
+
+## 1.13.3 - 2024-06-23
+
+### Bugs fixes 🐛
+
+* hotfix: bad default_time value type introduced in #300 by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/302>
+
+## 1.13.2 - 2024-06-23
+
+### Features and enhancements 🎉
+
+* Feature: add per-feed custom titles and abstracts by @toomanybrians in <https://github.com/Guts/mkdocs-rss-plugin/pull/295>
+* Improve: convert docstrings to google convention by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/298>
+* fix(config): do not raise error for bad date_from_meta.default_time value and fallback to default by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/300>
+
+### New Contributors
+
+* @toomanybrians made their first contribution in <https://github.com/Guts/mkdocs-rss-plugin/pull/295>
+
+## 1.13.1 - 2024-06-15
+
+### Bugs fixes 🐛
+
+* Fix remote image length warnings using requests instead of urllib (standard lib) by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/289>
+
+### Features and enhancements 🎉
+
+* refacto(cleanup): remove python 3.9 related code and deps (timezone) by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/288>
+* improve(logs): enhance some logs and type hints  by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/293>
+* Feature: improve social cards integration by @Guts in <https://github.com/Guts/mkdocs-rss-plugin/pull/294>
 
 ## 1.13.0 - 2024-06-10